+++ src/lib.rs
@@ -53,7 +53,7 @@ pub const CONSOLE: bool = false;
 // TODO: add documentation on allowed parameters for time
 pub const TIMESLOT_DURATION: u64 = 250;
 pub const CHALLENGE_LOOKBACK: u64 = 4;
-pub const TIMESLOTS_PER_EPOCH: u64 = 4;
+pub const TIMESLOTS_PER_EPOCH: usize = 4;
 pub const EPOCH_GRACE_PERIOD: Duration =
-    Duration::from_millis(TIMESLOTS_PER_EPOCH * TIMESLOT_DURATION);
+    Duration::from_millis(TIMESLOTS_PER_EPOCH as u64 * TIMESLOT_DURATION);
 pub const SOLUTION_RANGE: u64 = std::u64::MAX / PLOT_SIZE as u64 / 2;
